Update helper_models.py
Predict.py was stopping the predictions when one of the input images  had desired ratio less than 4 pixels. Now, the minimum size was reduced to 1 pixel and then if it is less than that, the image is resized without keeping the aspect ratio to desired size. This specific filename will be then nted out